Extract shared query helpers into query_helpers module
Move extract_sql_params, apply_filter, build_query_sig_params and
build_query_body out of query_tools into a new lower-layer
query_helpers module so schema_tools and cli/verify can share them
without crossing sibling tool-implementation modules. Updates
.importlinter and tach.toml to declare the new layer.
