+++ .importlinter
@@ -10,6 +10,7 @@ layers =
     mcp_tools_sql.cli
     mcp_tools_sql.server
     mcp_tools_sql.schema_tools | mcp_tools_sql.query_tools | mcp_tools_sql.update_tools | mcp_tools_sql.validation_tools
+    mcp_tools_sql.query_helpers
     mcp_tools_sql.tool_builder
     mcp_tools_sql.backends | mcp_tools_sql.formatting | mcp_tools_sql.tool_logging
     mcp_tools_sql.config
@@ -27,6 +28,7 @@ forbidden_modules =
     mcp_tools_sql.formatting
     mcp_tools_sql.schema_tools
     mcp_tools_sql.query_tools
+    mcp_tools_sql.query_helpers
     mcp_tools_sql.update_tools
     mcp_tools_sql.validation_tools
     mcp_tools_sql.server
